More VC7 compatibility fixes
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@1893 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/build.cpp
@@ -426,7 +426,7 @@ int CEXEBuild::add_string_main(const char *string, int process) // returns offse
       if (build_userlangstrings.find(cp, 0, &idx) < 0) idx = -1;
     }
     free(cp);
-    if (idx >= 0) return -(idx+1+(sizeof(common_strings)+sizeof(installer_strings))/sizeof(int));
+    if (idx >= 0) return -((int)(idx+1+(sizeof(common_strings)+sizeof(installer_strings))/sizeof(int)));
   }
 
   if (!process) return build_strlist.add(string,2);
@@ -449,7 +449,7 @@ int CEXEBuild::add_string_uninst(const char *string, int process) // returns off
       if (ubuild_userlangstrings.find(cp, 0, &idx) < 0) idx = -1;
     }
     free(cp);
-    if (idx >= 0) return -(idx+1+(sizeof(common_strings)+sizeof(uninstall_strings))/sizeof(int));
+    if (idx >= 0) return -((int)(idx+1+(sizeof(common_strings)+sizeof(uninstall_strings))/sizeof(int)));
   }
 
   if (!process) return ubuild_strlist.add(string,2);
